refactor: use existing system key types instead of custom ones
The mapping was a bit more elegant before -- now it has less
redundancy (it's not so declarative which key types are which)
and it's no longer self contained.

But... the old version became less and less predictable with
every Ubuntu upgrade.  This version uses the same approach as
Neo (which is bundled with the system package), so hopefully
it will be a bit more stable.